backend/requirements.txt
Django~=4.1.6
djangorestframework~=3.14.0
django-filter~=22.1
django-cors-headers~=3.14.0
drf-yasg~=1.21.5
openpyxl~=3.1.1